v7.3 release
Changelog:
Constant SCVs implemented
Constant Probes reverted to idle nexus penalty
Scouting worker implemented for Zerg & Terran (Protoss already implemented)
Battlecruiser requirement fix
Ghost requirement fix
Nitro Packs requirement fix
Supply Depot requirement fix

+++ trunk/SCBuildOrderGUIView.cpp
@@ -436,7 +436,7 @@ void CSCBuildOrderGUIView::OnTimer(UINT TimerVal)
 
 		output->GetScrollRange(SB_VERT, &scrollMin, &scrollMax);
 		//output->SetScrollPos(SB_VERT, (int)(scrollPercent * (scrollMax - scrollMin)) + scrollMin);
-		output->LineScroll(output->GetLineCount() * scrollPercent);
+		output->LineScroll((int)(output->GetLineCount() * scrollPercent));
 	}
 }
